Tandem results attributes

Tandem results attributes

Slots

U_i_g0

matrix. Initial object membership function matrix.

B_j_q0

matrix. Initial factor/component matrix for the variables.

C_k_r0

matrix. Initial factor/component matrix for the occasions.

U_i_g

matrix. Final/updated object membership function matrix.

B_j_q

matrix. Final/updated factor/component matrix for the variables.

C_k_r

matrix. Final/updated factor/component matrix for the occasions.

Y_g_qr

matrix. Derived centroids in the reduced space (data matrix).

X_i_jk_scaled

matrix. Standardized dataset matrix.

BestTimeElapsed

numeric. Execution time for the best iterate.

BestLoop

numeric. Loop that obtained the best iterate.

BestKmIteration

numeric. Number of iteration until best iterate for the K-means.

BestFaIteration

numeric. Number of iteration until best iterate for the FA.

FaConverged

numeric. Flag to check if algorithm converged for the K-means.

KmConverged

numeric. Flag to check if algorithm converged for the Factor Decomposition.

nKmConverges

numeric. Number of loops that converged for the K-means.

nFaConverges

numeric. Number of loops that converged for the Factor decomposition.

TSS_full

numeric. Total deviance in the full-space.

BSS_full

numeric. Between deviance in the reduced-space.

RSS_full

numeric. Residual deviance in the reduced-space.

PF_full

numeric. PseudoF in the full-space.

TSS_reduced

numeric. Total deviance in the reduced-space.

BSS_reduced

numeric. Between deviance in the reduced-space.

RSS_reduced

numeric. Residual deviance in the reduced-space.

PF_reduced

numeric. PseudoF in the reduced-space.

PF

numeric. Actual PseudoF value to obtain best loop.

Labels

integer. Object cluster assignments.

FsKM

numeric. Objective function values for the KM best iterate.

FsFA

numeric. Objective function values for the FA best iterate.

Enorm

numeric. Average l2 norm of the residual norm.