T3Clus Model

Implements simultaneous version of TWCFTA

fit.t3clus(model, X_i_jk, full_tensor_shape, reduced_tensor_shape)

# S4 method for simultaneous
fit.t3clus(model, X_i_jk, full_tensor_shape, reduced_tensor_shape)

Arguments

model

Initialized simultaneous model.

X_i_jk

Matricized tensor along mode-1 (I objects).

full_tensor_shape

Dimensions of the tensor in full-space.

reduced_tensor_shape

Dimensions of tensor in the reduced-space.

Value

Output attributes accessible via the '@' operator.

U_i_g0 - Initial object membership function matrix

B_j_q0 - Initial factor/component matrix for the variables

C_k_r0 - Initial factor/component matrix for the occasions

U_i_g - Final/updated object membership function matrix

B_j_q - Final/updated factor/component matrix for the variables

C_k_r - Final/updated factor/component matrix for the occasions

Y_g_qr - Derived centroids in the reduced space (data matrix)

X_i_jk_scaled - Standardized dataset matrix

BestTimeElapsed - Execution time for the best iterate

BestLoop - Loop that obtained the best iterate

BestIteration - Iteration yielding the best results

Converged - Flag to check if algorithm converged for the K-means

nConverges - Number of loops that converged for the K-means

TSS_full - Total deviance in the full-space

BSS_full - Between deviance in the reduced-space

RSS_full - Residual deviance in the reduced-space

PF_full - PseudoF in the full-space

TSS_reduced - Total deviance in the reduced-space

BSS_reduced - Between deviance in the reduced-space

RSS_reduced - Residual deviance in the reduced-space

PF_reduced - PseudoF in the reduced-space

PF - Weighted PseudoF score

Labels - Object cluster assignments

Fs - Objective function values for the KM best iterate

Enorm - Average l2 norm of the residual norm.

Details

The procedure performs simultaneously the sequential TWCFTA model. The model finds B_j_q and C_k_r such that the between-clusters deviance of the component scores is maximized.

References

Tucker L (1966). “Some mathematical notes on three-mode factor analysis.” Psychometrika, 31(3), 279-311. doi:10.1007/BF02289464 , https://ideas.repec.org/a/spr/psycho/v31y1966i3p279-311.html. Rocci R, Vichi M (2005). “Three-Mode Component Analysis with Crisp or Fuzzy Partition of Units.” Psychometrika, 70, 715-736. doi:10.1007/s11336-001-0926-z . Vichi M, Rocci R, Kiers H (2007). “Simultaneous Component and Clustering Models for Three-way Data: Within and Between Approaches.” Journal of Classification, 24, 71-98. doi:10.1007/s00357-007-0006-x .

Examples

X_i_jk = generate_dataset()$X_i_jk
model = simultaneous()
t3clus = fit.t3clus(model, X_i_jk, c(8,5,4), c(3,3,2))
#> [1] "Loop"            "Best Iter"       "Loop Time"       "BSS Full (%)"   
#> [5] "BSS Reduced (%)" "PF Full"         "PF Reduced"      "Converged"      
#> [1]  1.0000  1.0000  0.0079 13.9500 27.9200  1.0144 11.1582  0.0000
#> [1]  2.0000  1.0000  0.0079 18.1400 36.3000  1.3189 14.5080  0.0000
#> [1]  3.0000  1.0000  0.0080 13.3900 26.8000  0.9736 10.7091  0.0000
#> [1]  4.0000  1.0000  0.0077 14.2800 61.2100  1.0383 11.4216  0.0000
#> [1]  5.0000  1.0000  0.0079 15.4300 64.9000  1.1222 12.3446  0.0000
#> [1]  6.0000  1.0000  0.0079 15.8400 31.7000  1.1518 12.6703  0.0000
#> [1]  7.0000  2.0000  0.0126 17.6000 64.0000  1.2797 14.0765  0.0000
#> [1]  8.0000  1.0000  0.0081 21.5500 43.1400  1.5672 17.2396  0.0000
#> [1]  9.0000  3.0000  0.0088 18.6900 72.3400  1.3595 14.9542  0.0000
#> [1] 10.0000  1.0000  0.0086 18.0000 36.0400  1.3094 14.4039  0.0000