TWCFTA model

Implements K-means clustering and afterwards factorial reduction in a sequential fashion.

fit.twcfta(model, X_i_jk, full_tensor_shape, reduced_tensor_shape)

# S4 method for tandem
fit.twcfta(model, X_i_jk, full_tensor_shape, reduced_tensor_shape)

Arguments

model

Initialized tandem model.

X_i_jk

Matricized tensor along mode-1 (I objects).

full_tensor_shape

Dimensions of the tensor in full space.

reduced_tensor_shape

Dimensions of tensor in the reduced space.

Value

Output attributes accessible via the '@' operator.

U_i_g0 - Initial object membership function matrix.

B_j_q0 - Initial factor/component matrix for the variables.

C_k_r0 - Initial factor/component matrix for the occasions.

U_i_g - Final/updated object membership function matrix.

B_j_q - Final/updated factor/component matrix for the variables.

C_k_r - Final/updated factor/component matrix for the occasions.

Y_g_qr - Derived centroids in the reduced space (data matrix).

X_i_jk_scaled - Standardized dataset matrix.

BestTimeElapsed - Execution time for the best iterate.

BestLoop - Loop that obtained the best iterate.

BestKmIteration - Number of iteration until best iterate for the K-means.

BestFaIteration - Number of iteration until best iterate for the FA.

FaConverged - Flag to check if algorithm converged for the K-means.

KmConverged - Flag to check if algorithm converged for the Factor Decomposition.

nKmConverges - Number of loops that converged for the K-means.

nFaConverges - Number of loops that converged for the Factor decomposition.

TSS_full - Total deviance in the full-space.

BSS_full - Between deviance in the reduced-space.

RSS_full - Residual deviance in the reduced-space.

PF_full - PseudoF in the full-space.

TSS_reduced - Total deviance in the reduced-space.

BSS_reduced - Between deviance in the reduced-space.

RSS_reduced - Residual deviance in the reduced-space.

PF_reduced - PseudoF in the reduced-space.

PF - Actual PseudoF value to obtain best loop.

Labels - Object cluster assignments.

FsKM - Objective function values for the KM best iterate.

FsFA - Objective function values for the FA best iterate.

Enorm - Average l2 norm of the residual norm.

Details

The procedure requires sequential clustering and factorial decomposition.

The K-means clustering algorithm is initially applied to the matricized tensor X_i_jk to obtain the centroids matrix X_g_jk and the membership matrix U_i_g.

The Tucker2 decomposition technique is then implemented on the centroids matrix X_g_jk to yield the core centroids matrix Y_g_qr and the component weights matrices B_j_q and C_k_r.

Note

This procedure is useful to further interpret the between clusters variability of the data and to understand the variables and/or occasions that most contribute to discriminate the clusters. However, the application of this technique could lead to the masking of variables that are not informative of the clustering structure.

since the Tucker2 model is applied after the clustering, this cannot help select the most relevant information for the clustering in the dataset.

References

Arabie P, Hubert L (1996). “Advances in Cluster Analysis Relevant to Marketing Research.” In Gaul W, Pfeifer D (eds.), From Data to Knowledge, 3--19. Tucker L (1966). “Some mathematical notes on three-mode factor analysis.” Psychometrika, 31(3), 279-311. doi:10.1007/BF02289464 , https://ideas.repec.org/a/spr/psycho/v31y1966i3p279-311.html.

See also

Examples

X_i_jk = generate_dataset()$X_i_jk
model = tandem()
twcfta = fit.twcfta(model, X_i_jk, c(8,5,4), c(3,3,2))
#>  [1] "Loop"            "Best KM Iter"    "Best FA Iter"    "Loop Time"      
#>  [5] "BSS Full (%)"    "BSS Reduced (%)" "PF Full"         "PF Reduced"     
#>  [9] "KM Converged"    "FA Converged"   
#>  [1]  1.0000  1.0000  8.0000  0.0022 23.1400 48.6300  1.6829 18.5123  1.0000
#> [10]  1.0000
#>  [1]  2.0000  1.0000  3.0000  0.0013 20.5900 42.0200  1.4978 16.4756  1.0000
#> [10]  1.0000
#>  [1]  3.0000  1.0000  8.0000  0.0023 15.6900 40.0700  1.1414 12.5553  1.0000
#> [10]  1.0000
#>  [1]  4.0000  2.0000  8.0000  0.0026 27.6700 57.0300  2.0126 22.1382  1.0000
#> [10]  1.0000
#>  [1]  5.0000  2.0000  4.0000  0.0019 26.8800 55.3300  1.9546 21.5004  1.0000
#> [10]  1.0000
#>  [1]  6.0000  2.0000  4.0000  0.0020 32.9800 63.5600  2.3988 26.3872  1.0000
#> [10]  1.0000
#>  [1]  7.0000  1.0000  4.0000  0.0016 20.3700 46.3400  1.4813 16.2945  1.0000
#> [10]  1.0000
#>  [1]  8.0000  1.0000  4.0000  0.0015 19.0300 51.3300  1.3843 15.2268  1.0000
#> [10]  1.0000
#>  [1]  9.0000  1.0000  4.0000  0.0014 16.3100 45.9200  1.1862 13.0477  1.0000
#> [10]  1.0000
#>  [1] 10.0000  2.0000  4.0000  0.0018 25.0600 56.6300  1.8222 20.0442  1.0000
#> [10]  1.0000