TWFCTA model

Implements factorial reduction and then K-means clustering in a sequential fashion.

fit.twfcta(model, X_i_jk, full_tensor_shape, reduced_tensor_shape)

# S4 method for tandem
fit.twfcta(model, X_i_jk, full_tensor_shape, reduced_tensor_shape)

Arguments

model

Initialized tandem model.

X_i_jk

Matricized tensor along mode-1 (I objects).

full_tensor_shape

Dimensions of the tensor in full space.

reduced_tensor_shape

Dimensions of tensor in the reduced space.

Value

Output attributes accesible via the '@' operator.

U_i_g0 - Initial object membership function matrix.

B_j_q0 - Initial factor/component matrix for the variables.

C_k_r0 - Initial factor/component matrix for the occasions.

U_i_g - Final/updated object membership function matrix.

B_j_q - Final/updated factor/component matrix for the variables.

C_k_r - Final/updated factor/component matrix for the occasions.

Y_g_qr - Derived centroids in the reduced space (data matrix).

X_i_jk_scaled - Standardized dataset matrix.

BestTimeElapsed - Execution time for the best iterate.

BestLoop - Loop that obtained the best iterate.

BestKmIteration - Number of iteration until best iterate for the K-means.

BestFaIteration - Number of iteration until best iterate for the FA.

FaConverged - Flag to check if algorithm converged for the K-means.

KmConverged - Flag to check if algorithm converged for the Factor Decomposition.

nKmConverges - Number of loops that converged for the K-means.

nFaConverges - Number of loops that converged for the Factor decomposition.

TSS_full - Total deviance in the full-space.

BSS_full - Between deviance in the reduced-space.

RSS_full - Residual deviance in the reduced-space.

PF_full - PseudoF in the full-space.

TSS_reduced - Total deviance in the reduced-space.

BSS_reduced - Between deviance in the reduced-space.

RSS_reduced - Residual deviance in the reduced-space.

PF_reduced - PseudoF in the reduced-space.

PF - Actual PseudoF value to obtain best loop.

Labels - Object cluster assignments.

FsKM - Objective function values for the KM best iterate.

FsFA - Objective function values for the FA best iterate.

Enorm - Average l2 norm of the residual norm.

Details

The procedure implements sequential factorial decomposition and clustering.

The technique performs Tucker2 decomposition on the X_i_jk matrix to obtain the matrix of component scores Y_i_qr with component weights matrices B_j_q and C_k_r.

The K-means clustering algorithm is then applied to the component scores matrix Y_i_qr to obtain the desired core centroids matrix Y_g_qr and its associated stochastic membership function matrix U_i_g.

Note

The technique helps interpret the within clusters variability of the data. The Tucker2 tends to explain most of the total variation in the dataset. Hence, the variance of variables that do not contribute to the clustering structure in the dataset is also included.

The Tucker2 dimensions may still mask some essential clustering structures in the dataset.

References

Arabie P, Hubert L (1996). “Advances in Cluster Analysis Relevant to Marketing Research.” In Gaul W, Pfeifer D (eds.), From Data to Knowledge, 3--19. Tucker L (1966). “Some mathematical notes on three-mode factor analysis.” Psychometrika, 31(3), 279-311. doi:10.1007/BF02289464 , https://ideas.repec.org/a/spr/psycho/v31y1966i3p279-311.html.

See also

Examples

X_i_jk = generate_dataset()$X_i_jk
model = tandem()
twfCta = fit.twfcta(model, X_i_jk, c(8,5,4), c(3,3,2))
#>  [1] "Loop"            "Best KM Iter"    "Best FA Iter"    "Loop Time"      
#>  [5] "BSS Full (%)"    "BSS Reduced (%)" "PF Full"         "PF Reduced"     
#>  [9] "KM Converged"    "FA Converged"   
#>  [1]  1.0000  2.0000 10.0000  0.0032 44.0900 44.0900  1.5988 17.5863  1.0000
#> [10]  0.0000
#>  [1]  2.0000  2.0000 10.0000  0.0032 51.2400 51.2400  1.8579 20.4371  1.0000
#> [10]  0.0000
#>  [1]  3.0000  1.0000 10.0000  0.0029 42.1000 42.1000  1.5265 16.7913  1.0000
#> [10]  0.0000
#>  [1]  4.0000  3.0000 10.0000  0.0037 63.6300 63.6300  2.3074 25.3811  1.0000
#> [10]  0.0000
#>  [1]  5.0000  2.0000 10.0000  0.0033 52.6700 52.6700  1.9098 21.0082  1.0000
#> [10]  0.0000
#>  [1]  6.0000  2.0000 10.0000  0.0032 46.7100 46.7100  1.6936 18.6300  1.0000
#> [10]  0.0000
#>  [1]  7.0000  2.0000 10.0000  0.0032 63.6300 63.6300  2.3074 25.3811  1.0000
#> [10]  0.0000
#>  [1]  8.0000  2.0000 10.0000  0.0034 41.8000 41.8000  1.5158 16.6733  1.0000
#> [10]  0.0000
#>  [1]  9.0000  1.0000 10.0000  0.0031 44.0900 44.0900  1.5988 17.5863  1.0000
#> [10]  0.0000
#>  [1] 10.0000  3.0000 10.0000  0.0038 60.4100 60.4100  2.1905 24.0959  1.0000
#> [10]  0.0000