Three-Mode Dataset Generator for Simulations

Generate G clustered synthetic dataset of I objects measured on J variables for K occasions with additive noise.

generate_dataset(
  I = 8,
  J = 5,
  K = 4,
  G = 3,
  Q = 3,
  R = 2,
  centroids_spread = c(0, 1),
  noise_mean = 0,
  noise_stdev = 0.5,
  seed = NULL
)

Arguments

I

Number of objects.

J

Number of variables per occasion.

K

Number of occasions.

G

Number of clusters.

Q

Number of factors for the variables.

R

Number of factors for the occasions.

centroids_spread

interval from which to uniformly pick the centroids.

noise_mean

Mean of noise to generate.

noise_stdev

Noise effect level/spread/standard deviation.

seed

Seed for random sequence generation.

Value

Z_i_jk: Component scores in the full space.

E_i_jk: Generated noise at the given noise level.

X_i_jk: Dataset with noise level set to noise_stdev specified.

Y_g_qr: Centroids matrix in the reduced space.

U_i_g: Stochastic membership function matrix.

B_j_q: Objects component scores matrix.

C_k_r: Occasions component scores matrix.

Examples

generate_dataset(seed=0)
#> $Z_i_jk
#>           [,1]      [,2]       [,3]       [,4]      [,5]      [,6]      [,7]
#> [1,] 0.2631313 0.1008410 0.33039890 0.33039890 0.1008410 0.1456460 0.1920519
#> [2,] 0.4050685 0.4491948 0.31455702 0.31455702 0.4491948 0.1248445 0.3849207
#> [3,] 0.6421999 0.4723376 0.03089314 0.03089314 0.4723376 0.4857985 0.2488496
#> [4,] 0.6421999 0.4723376 0.03089314 0.03089314 0.4723376 0.4857985 0.2488496
#> [5,] 0.4050685 0.4491948 0.31455702 0.31455702 0.4491948 0.1248445 0.3849207
#> [6,] 0.4050685 0.4491948 0.31455702 0.31455702 0.4491948 0.1248445 0.3849207
#> [7,] 0.4050685 0.4491948 0.31455702 0.31455702 0.4491948 0.1248445 0.3849207
#> [8,] 0.6421999 0.4723376 0.03089314 0.03089314 0.4723376 0.4857985 0.2488496
#>           [,8]      [,9]     [,10]     [,11]     [,12]     [,13]     [,14]
#> [1,] 0.3588093 0.3588093 0.1920519 0.1456460 0.1920519 0.3588093 0.3588093
#> [2,] 0.4959530 0.4959530 0.3849207 0.1248445 0.3849207 0.4959530 0.4959530
#> [3,] 0.1900176 0.1900176 0.2488496 0.4857985 0.2488496 0.1900176 0.1900176
#> [4,] 0.1900176 0.1900176 0.2488496 0.4857985 0.2488496 0.1900176 0.1900176
#> [5,] 0.4959530 0.4959530 0.3849207 0.1248445 0.3849207 0.4959530 0.4959530
#> [6,] 0.4959530 0.4959530 0.3849207 0.1248445 0.3849207 0.4959530 0.4959530
#> [7,] 0.4959530 0.4959530 0.3849207 0.1248445 0.3849207 0.4959530 0.4959530
#> [8,] 0.1900176 0.1900176 0.2488496 0.4857985 0.2488496 0.1900176 0.1900176
#>          [,15]     [,16]     [,17]      [,18]      [,19]     [,20]
#> [1,] 0.1920519 0.2631313 0.1008410 0.33039890 0.33039890 0.1008410
#> [2,] 0.3849207 0.4050685 0.4491948 0.31455702 0.31455702 0.4491948
#> [3,] 0.2488496 0.6421999 0.4723376 0.03089314 0.03089314 0.4723376
#> [4,] 0.2488496 0.6421999 0.4723376 0.03089314 0.03089314 0.4723376
#> [5,] 0.3849207 0.4050685 0.4491948 0.31455702 0.31455702 0.4491948
#> [6,] 0.3849207 0.4050685 0.4491948 0.31455702 0.31455702 0.4491948
#> [7,] 0.3849207 0.4050685 0.4491948 0.31455702 0.31455702 0.4491948
#> [8,] 0.2488496 0.6421999 0.4723376 0.03089314 0.03089314 0.4723376
#> 
#> $E_i_jk
#>            [,1]        [,2]        [,3]       [,4]       [,5]        [,6]
#> [1,]  0.3817967  0.21784165  0.54288468  0.3633754 -0.2263920  0.13306868
#> [2,] -0.3995046 -0.61876921 -0.34547692  0.5759559 -0.4160216 -0.18835136
#> [3,] -0.5738285 -0.11213394 -0.64229968  0.4960802 -0.5832853  1.22068231
#> [4,] -0.1447308  0.18869782  0.02336309 -0.2147566 -0.5327953 -0.39766956
#> [5,] -0.1496076  0.06666818 -0.11785328  0.6191521 -0.7818910 -0.02743874
#> [6,] -0.2057554  0.40209475 -0.27144413 -0.1396731  0.5782685  0.12507066
#> [7,]  0.1261117 -0.02855339 -0.21665516  0.8789515  0.4160236  0.30912165
#> [8,] -0.4459606  0.25180399 -0.32473582  0.2803730 -0.1136643 -0.08631175
#>              [,7]         [,8]        [,9]      [,10]       [,11]       [,12]
#> [1,] -1.111950137 -0.712549197  0.33206785  0.3860711  0.64965615  0.11450980
#> [2,] -0.631807192  0.182970562  0.55048455 -0.1097578 -0.43663106  0.49827196
#> [3,]  0.179364448  0.124206324  0.07188574 -0.2124051  0.00418548  0.39092959
#> [4,] -0.005522739  0.032644091 -0.05887680 -0.2094900 -0.44043586 -0.38838831
#> [5,] -0.470324581  0.009578196 -0.45603418  0.4984934  0.29812951 -0.30799495
#> [6,] -0.057912661  0.128669189 -0.71879312 -0.1378890  0.05985882  0.02329015
#> [7,] -0.407484354 -0.324505039 -0.39854476  0.6280094 -0.14108694 -0.56519289
#> [8,]  0.121131740 -0.059584381  0.62704155  0.3233372  0.72799420  0.28835939
#>            [,13]       [,14]        [,15]      [,16]       [,17]       [,18]
#> [1,] -0.64037472 -0.84009136 -0.534346356  0.7558361 -0.43489144 -0.55986003
#> [2,]  0.81272365  0.52687543 -0.401606609 -0.2378491 -0.45534034 -0.22408712
#> [3,] -0.25034830 -0.55979955 -0.556882568  0.3989582  0.37063815  0.23586819
#> [4,]  0.83914860  0.16780860  0.790045842 -0.4870013  0.03425577 -0.59024534
#> [5,] -0.20625994  0.24739788  0.748909381  0.3446863 -0.16187538  0.73512850
#> [6,] -0.48614342  0.06902635  0.131322729 -0.4779196 -0.54325152 -0.65571030
#> [7,]  0.01269143 -0.05939601 -0.616450600 -0.6158535 -0.50796447 -0.04826246
#> [8,]  0.01373767  0.09884213 -0.001861767 -0.4784459 -0.38389509  1.18485995
#>             [,19]      [,20]
#> [1,]  0.445313238  1.3293290
#> [2,] -0.126091581  0.8401391
#> [3,] -0.432881877  0.3897920
#> [4,]  0.291293000  0.3566203
#> [5,] -0.006264673 -0.2714410
#> [6,] -0.187427381  0.4428892
#> [7,]  0.158942868 -0.1742973
#> [8,] -0.244402817 -0.5040273
#> 
#> $X_i_jk
#>             [,1]       [,2]        [,3]       [,4]        [,5]        [,6]
#> [1,] 0.644928063  0.3186826  0.87328358  0.6937743 -0.12555102  0.27871470
#> [2,] 0.005563873 -0.1695744 -0.03091990  0.8905129  0.03317319 -0.06350688
#> [3,] 0.068371382  0.3602037 -0.61140654  0.5269733 -0.11094764  1.70648083
#> [4,] 0.497469100  0.6610355  0.05425622 -0.1838634 -0.06045766  0.08812896
#> [5,] 0.255460939  0.5158630  0.19670374  0.9337091 -0.33269618  0.09740574
#> [6,] 0.199313081  0.8512896  0.04311289  0.1748839  1.02746334  0.24991514
#> [7,] 0.531180222  0.4206415  0.09790186  1.1935086  0.86521841  0.43396612
#> [8,] 0.196239323  0.7241416 -0.29384269  0.3112662  0.35867329  0.39948676
#>             [,7]       [,8]        [,9]      [,10]       [,11]       [,12]
#> [1,] -0.91989828 -0.3537399  0.69087710 0.57812295  0.79530217  0.30656165
#> [2,] -0.24688648  0.6789236  1.04643760 0.27516290 -0.31178658  0.88319267
#> [3,]  0.42821407  0.3142239  0.26190333 0.03644448  0.48998399  0.63977921
#> [4,]  0.24332688  0.2226617  0.13114079 0.03935957  0.04536265 -0.13953869
#> [5,] -0.08540387  0.5055312  0.03991886 0.88341414  0.42297399  0.07692576
#> [6,]  0.32700805  0.6246222 -0.22284007 0.24703170  0.18470330  0.40821086
#> [7,] -0.02256364  0.1714480  0.09740828 1.01293012 -0.01624246 -0.18027218
#> [8,]  0.36998136  0.1304332  0.81705914 0.57218682  1.21379271  0.53720901
#>            [,13]      [,14]      [,15]       [,16]        [,17]      [,18]
#> [1,] -0.28156546 -0.4812821 -0.3422945  1.01896747 -0.334050471 -0.2294611
#> [2,]  1.30867670  1.0228285 -0.0166859  0.16721936 -0.006145499  0.0904699
#> [3,] -0.06033071 -0.3697820 -0.3080329  1.04115811  0.842975787  0.2667613
#> [4,]  1.02916619  0.3578262  1.0388955  0.15519861  0.506593401 -0.5593522
#> [5,]  0.28969310  0.7433509  1.1338301  0.74975485  0.287319465  1.0496855
#> [6,]  0.00980963  0.5649794  0.5162434 -0.07285105 -0.094056681 -0.3411533
#> [7,]  0.50864448  0.4365570 -0.2315299 -0.21078503 -0.058769631  0.2662946
#> [8,]  0.20375526  0.2888597  0.2469879  0.16375395  0.088442542  1.2157531
#>           [,19]       [,20]
#> [1,]  0.7757121  1.43016998
#> [2,]  0.1884654  1.28933394
#> [3,] -0.4019887  0.86212964
#> [4,]  0.3221861  0.82895789
#> [5,]  0.3082923  0.17775387
#> [6,]  0.1271296  0.89208403
#> [7,]  0.4734999  0.27489750
#> [8,] -0.2135097 -0.03168965
#> 
#> $Y_g_qr
#>           [,1]      [,2]       [,3]      [,4]      [,5]      [,6]
#> [1,] 0.3721239 0.2016819 0.66079779 0.2059746 0.3841037 0.7176185
#> [2,] 0.5728534 0.8983897 0.62911404 0.1765568 0.7698414 0.9919061
#> [3,] 0.9082078 0.9446753 0.06178627 0.6870228 0.4976992 0.3800352
#> 
#> $U_i_g
#>      [,1] [,2] [,3]
#> [1,]    1    0    0
#> [2,]    0    1    0
#> [3,]    0    0    1
#> [4,]    0    0    1
#> [5,]    0    1    0
#> [6,]    0    1    0
#> [7,]    0    1    0
#> [8,]    0    0    1
#> 
#> $B_j_q
#>      [,1]      [,2]      [,3]
#> [1,]    1 0.0000000 0.0000000
#> [2,]    0 0.7071068 0.0000000
#> [3,]    0 0.0000000 0.7071068
#> [4,]    0 0.0000000 0.7071068
#> [5,]    0 0.7071068 0.0000000
#> 
#> $C_k_r
#>           [,1]      [,2]
#> [1,] 0.7071068 0.0000000
#> [2,] 0.0000000 0.7071068
#> [3,] 0.0000000 0.7071068
#> [4,] 0.7071068 0.0000000
#> 
#> $U_labels
#> [1] 1 2 3 3 2 2 2 3
#> 
#> $B_labels
#> [1] 1 2 3 3 2
#> 
#> $C_labels
#> [1] 1 2 2 1
#>